+++ test/_setup/basic.sh
@@ -1,0 +1,2 @@
+rm ~/Downloads/Archive/DuckDB/scratch.duckdb
+duckdb ~/Downloads/Archive/DuckDB/scratch.duckdb < test/_setup/basic.sql