+++ test/html_test.dart
@@ -7,7 +7,21 @@ void tests() {
       final input = '''
 CREATE TABLE IF NOT EXISTS tmp (
     resource_type ENUM('GENERATOR','INTERTIE', 'LOAD') NOT NULL,
-    sch_bid_curve_type ENUM('BIDPRICE'),
+);
+''';
+      final generator = CodeGenerator(
+        input,
+        timezoneName: 'America/Los_Angeles',
+      );
+      var generateHtmlDocs = generator.generateHtmlDocs();
+      print(generateHtmlDocs);
+      // print(generateHtmlDocs);
+    });
+
+    test('generate Html for timestamptz', () {
+      final input = '''
+CREATE TABLE IF NOT EXISTS tmp (
+    hour_beginning TIMESTAMPTZ NOT NULL,
 );
 ''';
       final generator = CodeGenerator(