+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 TODO:
  - Implement required filters, that is filters without which the URL is invalid.
+ - Implement a QueryFilter class for Dart client
+ - Implement a method to go from a QueryFilter class to a URI fragment.
+
+## 0.1.3 (2026-01-04)
+- Improvements to both the Rust and Dart clients. 
 
 ## 0.1.2 (2026-01-01)
 - Started work the Dart code generation.  Progress on enums. 
-
 
 ## 0.1.1 (2025-12-30)
 - Improvements to the Rust API.